[Feature] 공연 결과 DB 저장

See merge request s15-webmobile1-sub1/S15P11D206!70

+++ backend/ssasukae/src/main/java/com/ssafy/ssasukae/domain/performance/dto/PerformanceResponseDTO.java
@@ -18,7 +18,7 @@ public class PerformanceResponseDTO {
         String title;
         String artist;
         String thumbnailUrl;
-        BigDecimal score;
+        Integer score;
         LocalDateTime performanceAt;
     }
 }
