+++ backend/ssasukae/src/main/java/com/ssafy/ssasukae/domain/performance/websocket/payload/LeaderboardItemPayload.java
@@ -1,7 +1,5 @@
 package com.ssafy.ssasukae.domain.performance.websocket.payload;
 
-import java.math.BigDecimal;
-
 // 리더보드 세부 요소 페이로드
 public record LeaderboardItemPayload(
     // 순위
@@ -11,4 +9,4 @@ public record LeaderboardItemPayload(
     String nickname,
     Long songId,
     String songTitle,
-    BigDecimal finalScore) {}
+    Integer finalScore) {}
